+++ src/main/java/com/shopping/visitor/ShoppingItemVisitor.java
@@ -1,6 +1,8 @@
 package com.shopping.visitor;
+import java.math.BigDecimal;
+
 import com.shopping.fruit.Fruit;
 
 public interface ShoppingItemVisitor {
-   float visit(Fruit fruit);
+   BigDecimal visit(Fruit fruit);
 }
